Cast frames-in-flight to int32_t
Cast the return value of m_logicalDevice->getMaxFramesInFlight() to int32_t when constructing the ImageReader. This ensures the argument matches the expected parameter type and avoids implicit conversion issues or build errors on stricter compilers.

+++ app/src/main/cpp/GraphicsEngine/source/components/assets/textures/CameraTexture.cpp
@@ -117,7 +117,7 @@ namespace ge {
       bestWidth, bestHeight,
       AIMAGE_FORMAT_YUV_420_888,
       AHARDWAREBUFFER_USAGE_GPU_SAMPLED_IMAGE,
-      m_logicalDevice->getMaxFramesInFlight(),
+      static_cast<int32_t>(m_logicalDevice->getMaxFramesInFlight()),
       &m_imageReader
     );
